feat: watch extension 세분화
