+++ Projects/App/Project.swift
@@ -46,7 +46,7 @@ let targets: [Target] = [
         implements: .WatchExtension,
         factory: .init(
             dependencies: [
-//                .watchShared
+                .watchShared
             ]
         )
     )
