+++ Projects/App/WatchExtension/Sources/Core/WatchConnectivityManager.swift
@@ -1,0 +1,8 @@
+//
+//  WatchConnectivityManager.swift
+//  AppWatchExtension
+//
+//  Created by 송영모 on 2023/06/26.
+//
+
+import Foundation
