+++ Projects/App/WatchExtension/Sources/Domain/HealthClient.swift
@@ -1,0 +1,9 @@
+//
+//  HealthClient.swift
+//  AppWatchExtension
+//
+//  Created by 송영모 on 2023/06/26.
+//
+
+import Foundation
+
